factor out attenuator_scan from macros
