+++ mfx/attenuator_scan.py
@@ -1,0 +1,68 @@
+def attenuator_scan_separate_runs(events=240, record=False, config=True, transmissions=[0.01,0.02,0.03]):
+    """
+    Runs through attenuator conditions and records each as an individual run
+
+    Parameters
+    ----------
+    events: int, optional
+        number of events. default 240
+
+    record: bool, optional
+        set True to record
+
+    transmissions: list of floats, optional
+        list of transmissions to run through. default [0.01,0.02,0.03]
+
+    Operations
+    ----------
+
+    """
+    from time import sleep
+    from mfx.db import daq, att, pp
+
+    pp.open()
+    for i in transmissions:
+        att(i)
+        time.sleep(3)
+        daq.begin(events=events,record=record,wait=True, use_l3t=False)
+        daq.end_run()
+    pp.close()
+    daq.disconnect()
+
+
+def attenuator_scan_single_run(events=240, record=False, transmissions=[0.01,0.02,0.03]):
+    """
+    Runs through attenuator conditions and records them all as one continuous run
+
+    Parameters
+    ----------
+    events: int, optional
+        number of events. default 240
+
+    record: bool, optional
+        set True to record
+
+    transmissions: list of floats, optional
+        list of transmissions to run through. default [0.01,0.02,0.03]
+
+    Operations
+    ----------
+
+    """
+    from time import sleep
+    from mfx.db import daq, att, pp
+
+    daq.end_run()
+    daq.disconnect()
+    try:
+        pp.open()
+        daq.configure(record=record)
+        time.sleep(3)
+        for i in transmissions:
+            att(i,wait=True)
+            time.sleep(3)
+            daq.begin(events=events,record=record,wait=True, use_l3t=False)
+    finally:
+        daq.end_run()
+        pp.close()
+        daq.disconnect()
